minimal worked project

+++ app/main.py
@@ -73,7 +73,6 @@ async def store_to_db(short_url: str, long_url: str):
             new_short_url = ShortUrl(url=long_url, short_code=short_url)
             session.add(new_short_url)
             session.flush()  # Чтобы получить ID до коммита
-
             # Отсоединяем объект от сессии, чтобы можно было использовать после закрытия
             session.expunge(new_short_url)
             context = {
@@ -87,7 +86,8 @@ async def store_to_db(short_url: str, long_url: str):
     except IntegrityError:
         return False
     except Exception as err:
-        raise ConnectionError
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Connection error with database')
 
 
 # Получение оригинальной ссылки
@@ -101,6 +101,10 @@ async def get_origin_url(short_url: str, request: Request):
         with get_session() as session:
             short_url_entity = session.scalar(
                 select(ShortUrl).where(ShortUrl.short_code==short_url))
+            if short_url_entity is None:
+                raise NotFoundError
+            short_url_entity.access_count += 1
+            session.add(short_url_entity)
             context = {
                 'id': short_url_entity.id,
                 'url': short_url_entity.url,
@@ -108,13 +112,13 @@ async def get_origin_url(short_url: str, request: Request):
                 'createdAt': short_url_entity.created_at.isoformat() + 'Z',
                 'updatedAt': short_url_entity.updated_at.isoformat() + 'Z',
             }
-    except AttributeError:
+    except NotFoundError:
         return JSONResponse(
             status_code=status.HTTP_404_NOT_FOUND,
             content='Short url not found')
     except Exception as err:
-        print(type(err))
-        raise ConnectionError
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Connection error with database')
 
     return JSONResponse(status_code=status.HTTP_200_OK, content=context)
 
@@ -126,6 +130,8 @@ async def update_origin_url(short_url: str, item: Item):
         with get_session() as session:
             short_url_entity = session.scalar(
                 select(ShortUrl).where(ShortUrl.short_code==short_url))
+            if short_url_entity is None:
+                raise NotFoundError
             short_url_entity.url = str(item.url)
             session.add(short_url_entity)
             context = {
@@ -135,15 +141,20 @@ async def update_origin_url(short_url: str, item: Item):
                 'createdAt': short_url_entity.created_at.isoformat() + 'Z',
                 'updatedAt': short_url_entity.updated_at.isoformat() + 'Z',
             }
-    except AttributeError:
+    except NotFoundError:
         return JSONResponse(
             status_code=status.HTTP_404_NOT_FOUND,
             content='Short url not found')
     except Exception as err:
-        raise ConnectionError
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Connection error with database')
 
     return JSONResponse(
         status_code=status.HTTP_200_OK, content=context)
+
+
+class NotFoundError(Exception):
+    pass
 
 
 # Удаление короткой ссылки
@@ -153,12 +164,43 @@ async def delete_short_url(short_url: str):
         with get_session() as session:
             short_url_entity = session.scalar(
                 select(ShortUrl).where(ShortUrl.short_code==short_url))
+            if short_url_entity is None:
+                raise NotFoundError
             session.delete(short_url_entity)
-    except AttributeError:
+    except NotFoundError:
         return JSONResponse(
             status_code=status.HTTP_404_NOT_FOUND,
             content='Short url not found')
     except Exception as err:
-        raise ConnectionError
+        print(type(err))
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Connection error with database')
 
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
+    return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content='')
+
+
+@app.get('/shorten/{short_url}/stats')
+async def get_stats(short_url: str):
+    try:
+        with get_session() as session:
+            short_url_entity = session.scalar(
+                select(ShortUrl).where(ShortUrl.short_code==short_url))
+            if short_url_entity is None:
+                raise NotFoundError
+            context = {
+                'id': short_url_entity.id,
+                'url': short_url_entity.url,
+                'shortCode': short_url_entity.short_code,
+                'createdAt': short_url_entity.created_at.isoformat() + 'Z',
+                'updatedAt': short_url_entity.updated_at.isoformat() + 'Z',
+                'accessCount': short_url_entity.access_count
+            }
+            return JSONResponse(
+                status_code=status.HTTP_200_OK, content=context)
+    except NotFoundError:
+        return JSONResponse(
+            status_code=status.HTTP_404_NOT_FOUND,
+            content='Short url not found')
+    except Exception as err:
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Connection error with database')
